+++ src/Symfony/Component/PasswordHasher/Hasher/PasswordHasherFactory.php
@@ -13,9 +13,6 @@ namespace Symfony\Component\PasswordHasher\Hasher;
 
 use Symfony\Component\PasswordHasher\Exception\LogicException;
 use Symfony\Component\PasswordHasher\PasswordHasherInterface;
-use Symfony\Component\Security\Core\Encoder\EncoderAwareInterface;
-use Symfony\Component\Security\Core\Encoder\PasswordEncoderInterface;
-use Symfony\Component\Security\Core\Encoder\PasswordHasherAdapter;
 use Symfony\Component\Security\Core\User\PasswordAuthenticatedUserInterface;
 
 /**
@@ -43,7 +40,7 @@ class PasswordHasherFactory implements PasswordHasherFactoryInterface
     {
         $hasherKey = null;
 
-        if (($user instanceof PasswordHasherAwareInterface && null !== $hasherName = $user->getPasswordHasherName()) || ($user instanceof EncoderAwareInterface && null !== $hasherName = $user->getEncoderName())) {
+        if (($user instanceof PasswordHasherAwareInterface && null !== $hasherName = $user->getPasswordHasherName())) {
             if (!\array_key_exists($hasherName, $this->passwordHashers)) {
                 throw new \RuntimeException(sprintf('The password hasher "%s" was not configured.', $hasherName));
             }
@@ -63,10 +60,7 @@ class PasswordHasherFactory implements PasswordHasherFactoryInterface
         }
 
         if (!$this->passwordHashers[$hasherKey] instanceof PasswordHasherInterface) {
-            $this->passwordHashers[$hasherKey] = $this->passwordHashers[$hasherKey] instanceof PasswordEncoderInterface
-                ? new PasswordHasherAdapter($this->passwordHashers[$hasherKey])
-                : $this->createHasher($this->passwordHashers[$hasherKey])
-            ;
+            $this->passwordHashers[$hasherKey] = $this->createHasher($this->passwordHashers[$hasherKey]);
         }
 
         return $this->passwordHashers[$hasherKey];
@@ -91,9 +85,6 @@ class PasswordHasherFactory implements PasswordHasherFactoryInterface
         }
 
         $hasher = new $config['class'](...$config['arguments']);
-        if (!$hasher instanceof PasswordHasherInterface && $hasher instanceof PasswordEncoderInterface) {
-            $hasher = new PasswordHasherAdapter($hasher);
-        }
 
         if ($isExtra || !\in_array($config['class'], [NativePasswordHasher::class, SodiumPasswordHasher::class], true)) {
             return $hasher;